Make LibWCL compile on Ubuntu 12.04, Ubuntu 14.04, and OSX.
Squashed commit of the following:

commit 9e155fbc6cef5256382884f9469eee927e9a0b71
Author: Michael Marner <michael.marner@unisa.edu.au>
Date:   Tue Jan 6 14:21:43 2015 +1030

    Ignore the .cc files generated by flex/yacc

commit 1c1d01d815a14a0e02923e5f19d6dda29d699a59
Author: Michael Marner <michael.marner@unisa.edu.au>
Date:   Tue Jan 6 11:28:32 2015 +1030

    Deal with yacc's files differently depending on AM version.

commit c7e511c51cbee8f0e1d990edf234fbd2843dfc3b
Author: Michael Marner <michael.marner@unisa.edu.au>
Date:   Tue Jan 6 11:27:54 2015 +1030

    Get inttypes.h to define the printf macros.

    * Needed to print uint64_t without getting compiler warnings.

commit 5908c4f74fb9259901993de7be51cc6c3154c80f
Author: Michael Marner <michael.marner@unisa.edu.au>
Date:   Tue Jan 6 11:26:17 2015 +1030

    Deal with Automake 1.11 using different names for yacc output.

     * Add a check specifically for Ubuntu 12.04.
     * Modify the built sources directive in Makefile.am depending on OS.

    Really we should be checking for the automake version, but I can't find
    a simple macro for doing so. Open to suggestions.

commit 6baae1bf09a31a7c5914226e52489fc8f820fe36
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Jan 5 09:00:41 2015 +1030

    Ignore .dirstamp that has started appearing.

commit 1e2e671e8b71b501a4438b3c43d3ac3eecb57e80
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Dec 22 10:32:55 2014 +1030

    Some conditionals to deal with different AVCodec versions.

    * Ubuntu 14.04 is older than MacPorts.
    * Conditional build to deal with the way frames are alocated

commit 8b3ca31621f874a03ce23e88503804ee10edbc8a
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 12:51:17 2014 +1030

    Explicitly cast enum to int in comparisons in Serial.

    * Gets rid of some clang warnings, not sure if it was actually necessary though...

commit f5885a5420a6431700f98cd2a832a0a619021cc9
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 11:41:56 2014 +1030

    Update libav calls to replace deprecated functions.

    * What is with libav and deprecating function between every release?

commit d27b2af1281a934118dfbd10bd6d87d712d02f4d
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 11:41:18 2014 +1030

    Fix and simplify Ray's intersect methods.

    * Though untested becuase no gtest. That's why I'm on a different branch right?

commit dc66306acf6480562074566eb08dbdfac66d4540
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 11:27:03 2014 +1030

    Fix assignment in LineSegment.intersect

commit 7dbae111878afec0dbcd200eb55cdf3e9df33a61
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 11:20:01 2014 +1030

    Actually return reference in OBJGeometry operator=

    * Seriously, why does GCC not throw a warning with this stuff?

commit 9f725660e9fee90e329278fac84f14468a807c10
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 11:15:17 2014 +1030

    Fix LineSegment.intersect(LineSegment)

    * Simplified the code path so there is only a single return statement

    * Gets rid of the compile warning about reaching end of non-void function

    * Untested, because I have no google-test on my mac.

commit 4c1ec8f5fdcc3e9d01cce7f51d20545c492096f2
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 10:58:35 2014 +1030

    Put a return statement in BoundingSphere.contains

    * Method doesn't work. See bug report:
      http://discourse.wearablecomputerlab.net/t/libwcl-boundingsphere-contains-not-implemented/80/1

commit 50884c288b0499dfd092a4745785c9881155231e
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 10:57:48 2014 +1030

    Fix compilation errors introduced by me a couple of commits ago.

    * NetworkException now uses a const Socket* throughout.

commit 8aefde160e86ab76411b93c31a6cf7b118ac6d89
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 10:03:09 2014 +1030

    Change framerate value from int to float in DC-wcl framerate conversion

    * This should never have worked! Any fractional framerates (29.97, etc) would
      have been truncated to the nearest int.

    * Again, probably an obscure enough bug that it was never found before, but
      unit testing should have picked this up (if there was any)

commit c4ec51a964c64b65948abc0099c57ead62eaf4cb
Author: Michael Marner <michael@20papercups.net>
Date:   Tue Dec 16 09:59:45 2014 +1030

    Set SocketException.s correctly in the constructor.

    * Never used the parameter `is` passed in to constructor.
    * This fixes a bug, but it's probably so obscure nobody has ever run into it!
    * Clang gives waaaaay better error messages and far more thorough warnings than gcc

commit 7bd1fce27d3e0f117e682064cd70066713b3a61f
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Dec 15 16:12:29 2014 +1030

    Change the way GreyCode checks isValidRowCol

    * This may have introduced a bug, will be able to check when running video wall
      calibration

    * Regardless, using the ternary operator for a simple boolean comparison and
      return is a bit stupid.

commit 7c9c8197586737cd582fff6d4f49895d5f7e5cdb
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Dec 15 16:11:14 2014 +1030

    Fix arguments to printf in OBJParser, remove warnings.

    * I LOVE HOW C++ DOESN'T SPECIFY THE SIZE OF DATATYPES!
    * WHO KNOWS WHETHER long long unsigned int IS THE SAME AS uint64_t?

commit 2a374fe903a46f896ea00ce7cf7a7fd22de8ee51
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Dec 15 16:10:09 2014 +1030

    Remove Format7 warning from DC1394 camera.

    * It's unlikely that we will **ever** support Format7, so why keep nagging?

commit 8ffe2463c64bd99ddc4d6d367644409a5d42eb80
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Dec 15 15:55:52 2014 +1030

    Change access modifier of parser to using.

     o Remove compiler warning.

commit a816ab5e34d2797734a5eb57eae4e36cf182d974
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Dec 15 15:52:10 2014 +1030

    Added 'using' keyword to members of TCPServer.

    * Apparently the access modifiers we did have are deprecated.

commit 296e3a6fd0e0b22f8fffe8722f4b1770a7e85c11
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Dec 15 15:47:35 2014 +1030

    Fix names of FLEX/YACC generated files.

    * LibWCL now compiles on Ubuntu 14.04LTS
    * This now matches what OSX does
    * Autotools must have decided to change the names of generated files.
    * GREAT!

commit 8bca686ff94098865b37dd3fc22f596742c4aed6
Author: Michael Marner <michael@20papercups.net>
Date:   Mon Dec 15 14:46:43 2014 +1030

    Update build scripts for new autotools version.

    * Should also fix warnings on OSX.

+++ src/wcl/camera/DC1394Camera.cpp
@@ -34,7 +34,6 @@
 
 #include "CameraException.h"
 #include "DC1394Camera.h"
-#warning "[Note]: DC1394Camera: Doesn't currently support format7"
 
 #define ARRAY_SIZE(x) (sizeof(x)/sizeof(x[0]))
 
@@ -54,7 +53,7 @@ namespace wcl {
     struct DC1394tolibWCLFPS
     {
 	dc1394framerate_t   dc1394fps;
-	unsigned		libwclfps;
+	float libwclfps;
     };
 
     static DC1394tolibWCLModes formatConversion[] =
@@ -97,14 +96,14 @@ namespace wcl {
 
     static DC1394tolibWCLFPS fpsConversion[] =
     {
-	{ DC1394_FRAMERATE_1_875, 1.875 },
-	{ DC1394_FRAMERATE_3_75, 3.75 },
-	{ DC1394_FRAMERATE_7_5, 7.5 },
-	{ DC1394_FRAMERATE_15, 15.0 },
-	{ DC1394_FRAMERATE_30, 30.0 },
-	{ DC1394_FRAMERATE_60, 60.0 },
-	{ DC1394_FRAMERATE_120, 120.0 },
-	{ DC1394_FRAMERATE_240, 240.0 },
+	{ DC1394_FRAMERATE_1_875, 1.875f },
+	{ DC1394_FRAMERATE_3_75, 3.75f },
+	{ DC1394_FRAMERATE_7_5, 7.5f },
+	{ DC1394_FRAMERATE_15, 15.0f },
+	{ DC1394_FRAMERATE_30, 30.0f },
+	{ DC1394_FRAMERATE_60, 60.0f },
+	{ DC1394_FRAMERATE_120, 120.0f },
+	{ DC1394_FRAMERATE_240, 240.0f },
     };
